Атъебаный! Наконец завершил нормальный парсин фильтров в жсон в обе стороны!!!

+++ source/ParamChecker/Configuration/Configuration.cs
@@ -3,53 +3,49 @@ using System.Reflection;
 using Autodesk.Revit.UI;
 using Newtonsoft.Json;
 using ParamChecker.Models.Filters;
-using ParamChecker.Services;
-using ParamChecker.ViewModels.Windows;
 
 namespace ParamChecker.Configuration;
 
 public class Configuration
 {
-    
-    private readonly CategoryService _categoryService;
-    
+    private const string DirectoryName = "ParamCheckerSettings";
+
     private static readonly string DllPath = Assembly.GetExecutingAssembly().Location;
     private static readonly string DllDirectory = Path.GetDirectoryName(DllPath);
 
-    private const string DirectoryName = "ПапкаСНастройками потом переименуй";
-    private string _configFilePath = Path.Combine(DllDirectory, DirectoryName, "config.json");
-    
-    
-    public FilterConfigViewModel FilterConfigViewModel { get; set; }
-   
-    
-    public Configuration(CategoryService categoryService)
+    private readonly string _configFilePath = Path.Combine(DllDirectory, DirectoryName, "config.json");
+
+
+    public Configuration()
     {
         var directoryName = DirectoryName;
         var configName = "config.json";
-        
+
         // path to dll
         var dllPath = Assembly.GetExecutingAssembly().Location;
         CreateDir(dllPath, directoryName);
-        
+
         // path to dll`s directory
         var dllDir = Path.GetDirectoryName(dllPath);
-        
+
         // path to cfg`s directory
         var pathCfg = Path.Combine(dllDir, directoryName, configName);
-        
+
         // path to cfg`s
-        
+
         var dirCfg = Path.Combine(dllDir, directoryName);
-        
+
         if (!File.Exists(pathCfg))
         {
             CreateEmptyJsonFile(dirCfg, configName);
-            FilterConfigViewModel = new FilterConfigViewModel(categoryService);
+            FilterConfigModel = new FilterConfigModel();
             SaveConfig();
         }
     }
-    
+
+
+    public FilterConfigModel FilterConfigModel { get; set; }
+
     private static void CreateDir(string dllPath, string directoryName)
     {
         var dllDirectory = Path.GetDirectoryName(dllPath);
@@ -59,25 +55,20 @@ public class Configuration
             var x = Directory.CreateDirectory(configDirectoryPath);
         }
     }
+
     private static void CreateEmptyJsonFile(string directoryPath, string fileName)
     {
         var filePath = Path.Combine(directoryPath, fileName);
-        
-        if (!File.Exists(filePath))
-        {
-            File.WriteAllText(filePath, "{}");
-        }
+
+        if (!File.Exists(filePath)) File.WriteAllText(filePath, "{}");
     }
-    
+
     public static T LoadConfig<T>(string filePath) where T : class
     {
         try
         {
-            if (!File.Exists(filePath))
-            {
-                throw new FileNotFoundException("Файл конфигурации не найден", filePath);
-            }
-            
+            if (!File.Exists(filePath)) throw new FileNotFoundException("Файл конфигурации не найден", filePath);
+
             var json = File.ReadAllText(filePath);
             var config = JsonConvert.DeserializeObject<T>(json);
 
@@ -99,7 +90,7 @@ public class Configuration
             return null;
         }
     }
-    
+
     public void SaveConfig()
     {
         try
@@ -111,5 +102,10 @@ public class Configuration
         {
             TaskDialog.Show("Error", $"Ошибка с сохранние конфигурации: {ex.Message}");
         }
+    }
+
+    public string GetPath()
+    {
+        return _configFilePath;
     }
 }
